+++ src/Potion.Service/Program.cs
@@ -3,6 +3,11 @@ using Potion.Service;
 var builder = Host.CreateDefaultBuilder(args)
     .UseWindowsService()
     .UseContentRoot(AppContext.BaseDirectory)
+    .UseServiceProviderFactory(new DefaultServiceProviderFactory(new ServiceProviderOptions
+    {
+        ValidateOnBuild = true,
+        ValidateScopes = true
+    }))
     .ConfigureWebHostDefaults(webBuilder =>
     {
         webBuilder.UseStartup<Startup>();